restriction site and sequence label: start implementing
For the restriction site annotation, we want to plot a small label just above the main sequence track, in order to label the restriction site. For this, we create an additional track, named sequence_label, just above the main sequence, to write those labels. We start creating the FeatureSequenceLabel class.